+++ test/run_all_versions.sh
@@ -287,6 +287,7 @@ SUITES=(
 	wal_envelope
 	write_fsst_compressed
 	write_minmax_fastpath
+	zonemap_boundaries
 	zonemap_cost
 	zonemap_estimate_sample)